Ready for OpenWhisk deploy but too large.
Prepared for IBM Cloud with official OpenWhisk Python 3.11 runtime, but it won't deploy because the zipped action code is too large. This is because the BigQuery library in the virtualenv is too large. Will need to solve this by building a custom Python image or implementing it in a different programming language.